requirements.txt
flask==3.0.3
yt-dlp==2024.12.13
openai-whisper==20240930
torch==2.2.2
